Make TN state global.
