+++ wd-example/src/pages/_app.js
@@ -7,6 +7,7 @@ import Head from 'next/head';
 import { useStore } from 'base/state';
 import { Layout as DefaultLayout } from 'ui/app';
 import Menu from 'ui/app/menu';
+import { useEffect }  from "react";
 
 //	Styles
 import 'styles/index.scss';
@@ -16,6 +17,71 @@ import "@tokenscript/token-negotiator/dist/theme/style.css"
 export default function App({ Component, pageProps }) {
 	const Layout = Component.Layout ?? DefaultLayout;
 	const api = useStore( s => s.api );
+	//const selectedTokens = useStore( s => s.selectedTokens );
+
+	useEffect(()=> {
+
+		const init = async () => {
+
+			const { Client } = (await import("@tokenscript/token-negotiator"));
+
+			window.negotiator = new Client({
+				type: 'active',
+				issuers: [
+					{
+						collectionID: 'rinkeby-punk',
+						contract: '0x88b48f654c30e99bc2e4a1559b4dcf1ad93fa656',
+						chain: 'rinkeby',
+						openSeaSlug: 'rinkeby-punk'
+					},
+					{
+						collectionID: 'women-tribe',
+						contract: '0x88b48f654c30e99bc2e4a1559b4dcf1ad93fa656',
+						chain: 'rinkeby',
+						openSeaSlug: 'stl-rnd-women-tribe-nfts'
+					},
+					{
+						collectionID: 'zed',
+						contract: '0x88b48f654c30e99bc2e4a1559b4dcf1ad93fa656',
+						chain: 'rinkeby',
+						openSeaSlug: 'stl-rnd-zed'
+					},
+					{
+						collectionID: 'stl-rnd-bayc-derivatives',
+						contract: '0x88b48f654c30e99bc2e4a1559b4dcf1ad93fa656',
+						chain: 'rinkeby',
+						openSeaSlug: 'stl-rnd-bayc-derivatives'
+					},
+
+					{
+						collectionID: 'stl-rnd-riot-racers',
+						contract: '0x88b48f654c30e99bc2e4a1559b4dcf1ad93fa656',
+						chain: 'rinkeby',
+						openSeaSlug: 'stl-rnd-riot-racers'
+					}
+				],
+				options: {
+					overlay: {
+						openingHeading: "Open a new world of discounts available with your tokens.",
+						issuerHeading: "Get discount with Ticket",
+						repeatAction: "try again",
+						theme: "light",
+						position: "bottom-right"
+					},
+					filters: {},
+				}
+			});
+
+			window.negotiator.negotiate();
+
+			window.negotiator.on("tokens-selected", (data) => {
+				api.setSelectedTokens({...data.selectedTokens});
+			});
+		};
+
+		if (document.getElementsByClassName("overlay-tn")[0].childElementCount === 0)
+			init();
+	});
 
 	return (
 		<>
@@ -28,6 +94,7 @@ export default function App({ Component, pageProps }) {
 				<Component { ...pageProps } />
 			</Layout>
 			<Menu.Region onClose={ () => api.closeMenuView() } />
+			<div className="overlay-tn"/>
 		</>
 	);
 };
